+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,23 +1,10 @@
 package roomescape.controller;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.ResponseBody;
-import roomescape.entity.Reservation;
 
 @Controller
 public class ReservationController {
-
-    private final List<Reservation> reservations = new ArrayList<>();
-    private final AtomicLong index = new AtomicLong(1L);
 
     @GetMapping("/admin")
     public String getMainPage() {
@@ -29,28 +16,4 @@ public class ReservationController {
         return "admin/reservation-legacy";
     }
 
-    @GetMapping("/reservations")
-    @ResponseBody
-    public List<Reservation> readReservation() {
-        return reservations;
-    }
-
-    @PostMapping("/reservations")
-    @ResponseBody
-    public Reservation postReservation(@RequestBody Reservation reservation) {
-        Reservation newReservation = Reservation.toEntity(reservation, index.getAndIncrement());
-        reservations.add(newReservation);
-        return newReservation;
-    }
-
-    @DeleteMapping("/reservations/{id}")
-    public String deleteReservation(@PathVariable Long id) {
-        Reservation reservation = reservations.stream()
-                .filter(r -> Objects.equals(r.getId(), id))
-                .findFirst()
-                .orElseThrow();
-        reservations.remove(reservation);
-
-        return "admin/reservation-legacy";
-    }
 }
